+++ src/Content.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2020-2023 Andreas Möller
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE.md file that was distributed with this source code.
+ *
+ * @see https://github.com/ergebnis/front-matter
+ */
+
+namespace Ergebnis\FrontMatter;
+
+/**
+ * @psalm-immutable
+ */
+final class Content
+{
+    private function __construct(private readonly string $value)
+    {
+    }
+
+    public static function fromString(string $value): self
+    {
+        return new self($value);
+    }
+
+    public function toString(): string
+    {
+        return $this->value;
+    }
+}