+++ CHANGELOG.md
@@ -13,6 +13,7 @@ For a full diff see [`2.1.0...main`][2.1.0...main].
 - Dropped support for PHP 8.0 ([#393]), by [@localheinz]
 - Renamed `Parsed::fromFrontMatterAndContent()` to `Parsed::create()` ([#397]), by [@localheinz]
 - Renamed `Content` to `BodyMatter` and `Parsed::content()` to `Parsed::bodyMatter()` ([#398]), by [@localheinz]
+- Extracted `Content` as a value object ([#399]), by [@localheinz]
 
 ## [`2.1.0`][2.1.0]
 
@@ -104,5 +105,6 @@ For a full diff see [`4e97e14...0.1.0`][4e97e14...0.1.0].
 [#393]: https://github.com/ergebnis/front-matter/pull/393
 [#397]: https://github.com/ergebnis/front-matter/pull/397
 [#398]: https://github.com/ergebnis/front-matter/pull/398
+[#399]: https://github.com/ergebnis/front-matter/pull/399
 
 [@localheinz]: https://github.com/localheinz